RR-828 - Refactor timeline service and mapping
