+++ server/routes/overview/timelineController.ts
@@ -1,5 +1,4 @@
 import { RequestHandler } from 'express'
-import filterTimelineEvents from '../timelineResolver'
 import TimelineView from './timelineView'
 import TimelineService from '../../services/timelineService'
 
@@ -10,8 +9,7 @@ export default class TimelineController {
     const { prisonNumber } = req.params
     const { prisonerSummary } = req.session
 
-    const allTimelineEvents = await this.timelineService.getTimeline(prisonNumber, req.user.token, req.user.username)
-    const timeline = filterTimelineEvents(allTimelineEvents)
+    const timeline = await this.timelineService.getTimeline(prisonNumber, req.user.token, req.user.username)
     const view = new TimelineView(prisonerSummary, timeline)
     res.render('pages/overview/index', { ...view.renderArgs })
   }
