+++ server/services/timelineService.test.ts
@@ -1,12 +1,10 @@
-import type { TimelineResponse } from 'educationAndWorkPlanApiClient'
 import type { Timeline } from 'viewModels'
 import moment from 'moment'
 import PrisonService from './prisonService'
 import TimelineService from './timelineService'
 import EducationAndWorkPlanClient from '../data/educationAndWorkPlanClient'
-import { toTimeline } from '../data/mappers/timelineMapper'
+import toTimeline from '../data/mappers/timelineMapper'
 import aValidTimelineResponse from '../testsupport/timelineResponseTestDataBuilder'
-import aValidTimeline from '../testsupport/timelineTestDataBuilder'
 
 jest.mock('../data/mappers/timelineMapper')
 jest.mock('./prisonService')
@@ -28,25 +26,6 @@ describe('timelineService', () => {
   const userToken = 'a-user-token'
 
   describe('getTimeline', () => {
-    it('should get all timeline events', async () => {
-      // Given
-      const timelineResponse: TimelineResponse = aValidTimelineResponse()
-      educationAndWorkPlanClient.getTimeline.mockResolvedValue(timelineResponse)
-
-      const timeline: Timeline = aValidTimeline()
-      mockedTimelineMapper.mockReturnValue(timeline)
-
-      // When
-      const actual = await timelineService.getTimeline(prisonNumber, userToken, username)
-
-      // Then
-      expect(actual).toEqual(timeline)
-      expect(educationAndWorkPlanClient.getTimeline).toHaveBeenCalledWith(prisonNumber, userToken)
-      expect(mockedTimelineMapper).toHaveBeenCalledWith(timelineResponse)
-    })
-  })
-
-  describe('lookup prison names', () => {
     it('should get timeline given prison name lookups for several different prisons', async () => {
       // Given
       const timelineResponse = aValidTimelineResponse()
