+++ server/testsupport/timelineResponseTestDataBuilder.ts
@@ -1,4 +1,4 @@
-import type { TimelineResponse } from 'educationAndWorkPlanApiClient'
+import type { TimelineResponse, TimelineEventResponse } from 'educationAndWorkPlanApiClient'
 
 export default function aValidTimelineResponse(): TimelineResponse {
   return {
@@ -15,7 +15,7 @@ export default function aValidTimelineResponse(): TimelineResponse {
         correlationId: '6457a634-6dbe-4179-983b-74e92883232c',
         contextualInfo: {},
         actionedByDisplayName: undefined,
-      },
+      } as TimelineEventResponse,
       {
         reference: 'cd98ea4c-b415-48d9-a600-9068cefe65e4x',
         sourceReference: '33bc1045-7368-47c4-a261-4d616b7b51b9',
@@ -26,7 +26,7 @@ export default function aValidTimelineResponse(): TimelineResponse {
         correlationId: '246aa049-c5df-459d-8231-bdeab3936d0f',
         contextualInfo: {},
         actionedByDisplayName: 'Ralph Gen',
-      },
+      } as TimelineEventResponse,
       {
         reference: '3f0423e5-200b-48c9-8414-f04e336897ff',
         sourceReference: 'f190e844-3aa1-4f04-81d5-6be2bf9721cc',
@@ -39,7 +39,7 @@ export default function aValidTimelineResponse(): TimelineResponse {
           GOAL_TITLE: 'Learn French',
         },
         actionedByDisplayName: 'Ralph Gen',
-      },
+      } as TimelineEventResponse,
       {
         reference: '0a5d6abf-3504-4bb1-840b-370e3f35cd44',
         sourceReference: 'f190e844-3aa1-4f04-81d5-6be2bf9721cc',
@@ -52,7 +52,7 @@ export default function aValidTimelineResponse(): TimelineResponse {
           GOAL_TITLE: 'Learn French',
         },
         actionedByDisplayName: 'Ralph Gen',
-      },
+      } as TimelineEventResponse,
       {
         reference: '2589e577-64cf-4050-aff1-c2a28086e452',
         sourceReference: '1211013',
@@ -65,7 +65,7 @@ export default function aValidTimelineResponse(): TimelineResponse {
           PRISON_TRANSFERRED_FROM: 'MDI',
         },
         actionedByDisplayName: undefined,
-      },
+      } as TimelineEventResponse,
       {
         reference: 'b046eccc-eadc-4351-8e2c-c5da12fca973',
         sourceReference: '93646ba2-d0f9-42a9-9f87-3eb985ab6fe9',
@@ -78,7 +78,7 @@ export default function aValidTimelineResponse(): TimelineResponse {
           GOAL_TITLE: 'Learn French',
         },
         actionedByDisplayName: 'Ralph Gen',
-      },
+      } as TimelineEventResponse,
       {
         reference: '12f45f57-72e8-484a-8325-137507214634',
         sourceReference: '1211013',
@@ -89,7 +89,7 @@ export default function aValidTimelineResponse(): TimelineResponse {
         correlationId: '7e82200d-6251-4d38-9207-1ce49650dedf',
         contextualInfo: {},
         actionedByDisplayName: undefined,
-      },
+      } as TimelineEventResponse,
       {
         reference: '8d16c91e-58d3-4fc6-9820-75ae7395eaf2',
         sourceReference: '1211014',
@@ -100,7 +100,7 @@ export default function aValidTimelineResponse(): TimelineResponse {
         correlationId: '94e445d5-e4fa-4e32-93f5-0e5d6e3dbb2d',
         contextualInfo: {},
         actionedByDisplayName: undefined,
-      },
+      } as TimelineEventResponse,
     ],
   }
 }
